+++ src/data/shared-page-props.tsx
@@ -4,12 +4,12 @@ import { Entity } from "src/domain/data";
 import { defaultLocale } from "src/locales/config";
 import {
   getCanton,
-  getDimensionValuesAndLabels,
-  getElectricityPriceCube,
   getMunicipality,
+  getMunicipalityOperators,
   getOperator,
   getOperatorMunicipalities,
 } from "src/rdf/queries";
+import { sparqlClient } from "src/rdf/sparql-client";
 
 export type PageParams = { locale: string; id: string; entity: Entity };
 
@@ -43,12 +43,15 @@ export type Props =
       status: "notfound";
     };
 
-export const handleMunicipalityEntity = async (
-  params: Omit<PageParams, "entity"> & { res: ServerResponse<IncomingMessage> }
+export const getMunicipalityPageProps = async (
+  params: Omit<PageParams, "entity"> & {
+    res: ServerResponse<IncomingMessage>;
+    years: string[];
+  }
 ): Promise<
   Extract<Props, { entity: "municipality" } | { status: "notfound" }>
 > => {
-  const { id, locale, res } = params!;
+  const { id, locale, res, years } = params!;
   const municipality = await getMunicipality({ id });
 
   if (!municipality) {
@@ -56,13 +59,7 @@ export const handleMunicipalityEntity = async (
     return { status: "notfound" };
   }
 
-  const cube = await getElectricityPriceCube();
-
-  const operators = await getDimensionValuesAndLabels({
-    cube,
-    dimensionKey: "operator",
-    filters: { municipality: [id] },
-  });
+  const operators = await getMunicipalityOperators(sparqlClient, id, years);
 
   return {
     entity: "municipality",
@@ -75,7 +72,7 @@ export const handleMunicipalityEntity = async (
     locale: locale ?? defaultLocale,
   };
 };
-export const handleOperatorsEntity = async (
+export const getOperatorsPageProps = async (
   params: Omit<PageParams, "entity"> & { res: ServerResponse<IncomingMessage> }
 ): Promise<Extract<Props, { entity: "operator" } | { status: "notfound" }>> => {
   const { id, locale, res } = params!;
@@ -97,7 +94,7 @@ export const handleOperatorsEntity = async (
   };
 };
 
-export const handleCantonEntity = async (
+export const getCantonPageProps = async (
   params: Omit<PageParams, "entity"> & { res: ServerResponse<IncomingMessage> }
 ): Promise<Extract<Props, { entity: "canton" } | { status: "notfound" }>> => {
   const { id, locale, res } = params!;