+++ src/domain/query-states.ts
@@ -135,7 +135,7 @@ export const sunshineDetailsLink = makeLinkGenerator(
 );
 
 const sunshineOverviewFiltersSchema = z.object({
-  year: z.string().default("2025"),
+  year: z.string().default(buildEnv.CURRENT_PERIOD),
   category: z.string().default("C2"),
   networkLevel: z.enum(["NE5", "NE6", "NE7"]).default("NE5"),
 });