+++ src/graphql/resolvers.ts
@@ -47,8 +47,12 @@ import { fetchOperatorInfo, search } from "src/rdf/search-queries";
 import { asElectricityCategory } from "src/domain/data";
 import { asNetworkLevel } from "src/domain/sunshine";
 import { last, sortBy } from "lodash";
-import { CoverageCacheManager } from "src/rdf/coverage-ratio";
+import {
+  COVERAGE_RATIO_THRESHOLD,
+  CoverageCacheManager,
+} from "src/rdf/coverage-ratio";
 import { sparqlClient } from "src/rdf/sparql-client";
+import { truthy } from "src/lib/truthy";
 
 const gfmSyntax = require("micromark-extension-gfm");
 const gfmHtml = require("micromark-extension-gfm/html");
@@ -68,11 +72,6 @@ const expectedCubeDimensions = [
   "https://energy.ld.admin.ch/elcom/electricityprice/dimension/product",
   "https://cube.link/observedBy",
 ];
-
-/**
- * Under this threshold, observations are not returned
- */
-const COVERAGE_RATIO_THRESHOLD = 0.25;
 
 const Query: QueryResolvers = {
   sunshineData: async (_parent, args, context) => {
@@ -225,14 +224,19 @@ const Query: QueryResolvers = {
       ...o,
     })) as ResolvedOperatorObservation[];
 
-    const years = filters?.period;
-    if (years && observationFields && "coverageRatio" in observationFields) {
+    const years = Array.from(
+      new Set(operatorObservations.map((x) => x.period).filter(truthy))
+    );
+    if (years) {
       const defaultNetworkLevel = "NE7";
       const coverageManager = new CoverageCacheManager(sparqlClient);
       await coverageManager.prepare(years);
       operatorObservations.forEach((x) => {
-        x.coverageRatio =
-          coverageManager.getCoverage(x, defaultNetworkLevel) ?? 1;
+        const coverageRatio = coverageManager.getCoverage(
+          x,
+          defaultNetworkLevel
+        );
+        x.coverageRatio = coverageRatio;
         return x;
       });
     }