+++ src/pages/sunshine/[entity]/[id]/costs-and-tariffs.tsx
@@ -29,7 +29,7 @@ import TableComparisonCard from "src/components/table-comparison-card";
 import { TariffsTrendCard } from "src/components/tariffs-trend-card";
 import {
   DataServiceProps,
-  handleOperatorsEntity,
+  getOperatorsPageProps,
   PageParams,
   Props as SharedPageProps,
 } from "src/data/shared-page-props";
@@ -60,7 +60,7 @@ import {
 } from "src/lib/sunshine-data-service-context";
 import { truthy } from "src/lib/truthy";
 import { defaultLocale } from "src/locales/config";
-
+import { makePageTitle } from "src/utils/page-title";
 type Props =
   | (Extract<SharedPageProps, { entity: "operator"; status: "found" }> & {
       costsAndTariffs: Omit<
@@ -86,7 +86,7 @@ export const getServerSideProps: GetServerSideProps<Props, PageParams> = async (
     };
   }
 
-  const operatorProps = await handleOperatorsEntity({
+  const operatorProps = await getOperatorsPageProps({
     id,
     locale: locale ?? defaultLocale,
     res,
@@ -687,10 +687,7 @@ const CostsAndTariffs = (props: Props) => {
 
   const { id, name, entity } = props;
 
-  const pageTitle = `${getLocalizedLabel({ id: entity })} ${name} – ${t({
-    id: "sunshine.costs-and-tariffs.title",
-    message: "Costs and Tariffs",
-  })}`;
+  const pageTitle = makePageTitle(name);
 
   const handleTabChange = (
     _: React.SyntheticEvent,
@@ -713,17 +710,7 @@ const CostsAndTariffs = (props: Props) => {
   const mainContent = (
     <>
       <Head>
-        <title>
-          {t({
-            id: "sunshine.costs-and-tariffs.title",
-            message: "Costs and Tariffs",
-          })}
-          {" - "}
-          {t({
-            id: "site.title",
-            message: "Electricity tariffs in Switzerland",
-          })}
-        </title>
+        <title>{pageTitle}</title>
       </Head>
       <DetailsPageHeader>
         <DetailsPageTitle>