+++ src/pages/sunshine/[entity]/[id]/operational-standards.tsx
@@ -25,7 +25,7 @@ import {
 import TableComparisonCard from "src/components/table-comparison-card";
 import {
   DataServiceProps,
-  handleOperatorsEntity,
+  getOperatorsPageProps,
   PageParams,
   Props as SharedPageProps,
 } from "src/data/shared-page-props";
@@ -37,6 +37,7 @@ import {
   getSunshineDataServiceInfo,
 } from "src/lib/sunshine-data-service-context";
 import { defaultLocale } from "src/locales/config";
+import { makePageTitle } from "src/utils/page-title";
 
 type Props =
   | (Extract<SharedPageProps, { entity: "operator"; status: "found" }> & {
@@ -59,7 +60,7 @@ export const getServerSideProps: GetServerSideProps<Props, PageParams> = async (
     };
   }
 
-  const operatorProps = await handleOperatorsEntity({
+  const operatorProps = await getOperatorsPageProps({
     id,
     locale: locale ?? defaultLocale,
     res,
@@ -367,15 +368,12 @@ const PowerStability = (props: Props) => {
     <>
       <Head>
         <title>
-          {t({
-            id: "sunshine.operational-standards.title",
-            message: "Operational Standards",
-          })}
-          {" - "}
-          {t({
-            id: "site.title",
-            message: "Electricity tariffs in Switzerland",
-          })}
+          {makePageTitle(
+            t({
+              id: "sunshine.operational-standards.title",
+              message: "Operational Standards",
+            })
+          )}
         </title>
       </Head>
       <DetailsPageHeader>