+++ src/utils/page-title.tsx
@@ -1,0 +1,8 @@
+import { t } from "@lingui/macro";
+
+export const makePageTitle = (pageTitle?: string) => {
+  return `${pageTitle ? `${pageTitle} - ` : ""}${t({
+    id: "site.title",
+    message: "Electricity tariffs in Switzerland",
+  })}`;
+};